Final Commit
Followed all steps for this project. The "ODIN" logo and the text is transparent with a background. I put the background a little bit above the center point which looks good with the two anime characters.  I put the form inside of a div and the form inputs are inside another div. The "create account" button is within the form div which is easier than to place it outside of my form div. Finally, I put the "log in" inside of href link with a hover pseudo class. The sign up form service is also accessible on mobile devices!